feat: Add TouchableOpacity to MusicCardSearch component
The code changes in the Search component add a TouchableOpacity component to the MusicCardSearch component. This allows users to click on the "Adicionar à Playlist" button to add a music to the playlist. The handleAddMusicToPlaylist function is updated to pass the correct music ID when adding a music to the playlist. This enhancement improves the user experience by providing an interactive way to add music to a playlist.

Closes #123